+++ ci-cfg.el
@@ -1,3 +1,0 @@
-;;; configuration on CI, loaded by all Emacs invocations
-
-(package-initialize)