+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+# 1.3.0
+
+- allow vterm
+
 - add a function and keybinding (`C-c C-l`) to list methods.
 
 - make `C-c C-d` work with symbols that have a module path